Fix pokemon disappearing from animation window
Every now and then, as the party information is repeatedly updated, a pokemon's decrypted data will be incorrect. This has usually been in the stats department. When that's wrong, the pokemon won't validate, so it will disappear from the animation window. I've created ways to limit the updates and to ignore updates that occur due to invalidated pokemon data.

Unfortunately, there is a lingering issue of repeated party switching causing the whole emulator to shut down. This doesn't seem to be an issue when it's only the Lead pokemon being animated, but it is in the full party